now migration is async
